requirements.txt
celery
redis
python-decouple
pandas
fastparquet
scikit-learn==0.24.2
